requirements.txt
pillow
pyserial
ephem